added staying text

+++ consumer.py
@@ -73,11 +73,12 @@ for message in consumer:
     
     if not line == previous_line:
         print(line)
-        with open('lyrics.txt','r+') as myfile:
-            data = myfile.read()
-            myfile.seek(0)
-            myfile.write(line)
-            myfile.truncate()
+        if line is not "":
+            with open('lyrics.txt','r+') as myfile:
+                data = myfile.read()
+                myfile.seek(0)
+                myfile.write(line)
+                myfile.truncate()
 
         previous_line = line
